fix: 빌드오류 수정 및 폴더정리
